% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{add_ci_2plot}
\alias{add_ci_2plot}
\title{Add CI to Pretty Screeplot}
\usage{
add_ci_2plot(
  plot,
  CIs,
  met = "normal",
  color = "darkseagreen3",
  outline = "black",
  transparency = 0.5,
  type = "band",
  lt = "dotted"
)
}
\arguments{
\item{plot}{The plot object as obtained by pretty_scree()}

\item{CIs}{Data.frame obtained from getCIs() and booted_eigenvalues()}

\item{met}{the type of CI to be used default "normal"
| other options: "basic", "stud", "perc", "bca"}

\item{color}{If type of visual display of the CI = "band"
the color of the area between Upper and Lower}

\item{outline}{Color of the Outline for the Area}

\item{transparency}{Alpha value, how transparent shall the filling color be}

\item{type}{either "band" or "errorbars", defaults to "band"}

\item{lt}{Linetype, 0 for no outline around the band, other options include: "dotted", "dashed", "solid"}
}
\value{
A new pretty_plot with the Parallel Analysis,
including CIs around the observed Eigenvalues.
}
\description{
After CIs are determined by a combination of booted_eigenvalues() & getCIs()
This function provided with the plot object and the data.frame with CIs
adds CI around the observd eigenvalues
}
\author{
Björn Büdenbender
}
