% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{apa_factorLoadings}
\alias{apa_factorLoadings}
\title{Create an APA Formatted Table of Factor Loadings}
\usage{
apa_factorLoadings(
  fa_object,
  filepath = NA,
  table_caption = c("Table X", "PCA / EFA Factor Loadings"),
  overwrite = FALSE,
  nod = c(3, 2),
  n.obs = NA,
  cutoff = 0.3,
  ...
)
}
\arguments{
\item{fa_object}{Either an \code{\link[psych]{fa}} or \code{\link[psych]{principal}}, from
which the factor loadings are to be extracted}

\item{filepath}{(Optional) Path and filename were the APA ready table should
be saved, options include the common filetypes .docx (Word), .pptx (Powerpoint),
.html (Webpage). Default is filepath = NA. If NA is given, no file will be saved.}

\item{table_caption}{Takes a character vector. Recommend are the following elements
\itemize{
\item Table + number (e.g., "\strong{Table 3}")
\item The description, use APA capital case
(e.g., "\emph{PCA / EFA Factor Loadings}")
}}

\item{overwrite}{(Optional) Boolean, default is FALSE. When overwrite is
FALSE and the files already exists, a serialized version of the filename
will be created (i.e., appending _001, or _002). Utilizes \code{\link{serialNext}}}

\item{nod}{(Optional) Integer or Integer Vector. Number of Decimals.
In case of -1 a simple convention based on sample size is applied for determination
of number of decimal points. See ?datscience::get_number_of_decimals.
You can also provide an Integer vector, if you want different number of decimals
for the correlations and the summary stats. The first integer determines nod for
factor loadings, the second for additional stats. E.g., c(2,-1) would give 2 decimals
for factor loadings and apply the convention for Complexity, Uniqueness and Communality.
Default is nod = c(3,2). \cr
See \code{\link{get_number_of_decimals}} or\code{?datscience::get_number_of_decimals}}

\item{n.obs}{(Optional), Only required if you want to apply the convention to number
of decimals, and did calculate the factor analysis (PCA, EFA) directly on a correlation
or covariance matrix.}

\item{cutoff}{(Optional), Integer, determines the minimum threshold, values below are omitted.
Default is cutoff = 0.3.}

\item{...}{(Optional), Additional arguments that can be passed to \code{\link{format_flextable}}
(e.g., fontsize, font ...) or to \code{\link{serialNext}}}
}
\value{
A \code{\link[flextable]{flextable}} object with APA ready correlation table. If a filepath is provided
it also creates the respective file (e.g., a word .docx file)
}
\description{
Extracts factor loadings from factor analyses (either PCA or EFA) conducthed
with the \href{https://cran.r-project.org/web/packages/psychTools/vignettes/factor.pdf}{psych package},
and returns a APA 7th formatted table with the factor loadings and additional metrics (Communality, Uniqueness, Complexity).
If desired this table can directly be exported to a Word file (.docx).
}
\examples{
datscience::apa_factorLoadings(psych::fa(mtcars, nfactors = 2))
}
\seealso{
\code{\link{get_number_of_decimals}}, \code{\link{format_flextable}},
\code{\link{serialNext}}
}
\author{
Bjoern Buedenbender
}
