% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{boxplot_t_test}
\alias{boxplot_t_test}
\title{Creates Boxplots with Significance Makers}
\usage{
boxplot_t_test(df, dependentvars, group, adjust_p = "BH", ylimits = c(0, 150))
}
\arguments{
\item{df}{data.frame.}

\item{dependentvars}{Character vector.}

\item{group}{Character vector.}

\item{adjust_p}{Character vector.}

\item{ylimits}{Numeric vector.}
}
\value{
List(Plot and stats)
}
\description{
Based on the following: Guide https://www.datanovia.com/en/blog/how-to-perform-multiple-t-test-in-r-for-different-variables/
Given a vector of dependent variables (DV), creates boxplots
with facetwrap for all DVs and calculates an independent sample T-Test
to include significance bars
}
\examples{
 \dontrun{boxplot_t_test(mtcars,c("mpg","hp"),group="am")}

}
