% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{citations_appendix}
\alias{citations_appendix}
\title{R Package Citations}
\usage{
citations_appendix(outdirectory = "./")
}
\arguments{
\item{outputdirectory}{A character vector for the output directory (for the two files, .bib and .csv). Default ist ./}
}
\value{
Two Files for a Package List and Citations for the Appendix of the Paper
}
\description{
Function to create a package table with the following columns: Packagename | Version | Maintainer | Citation
}
\examples{
citations_appendix()

}
\author{
Björn Büdenbender
}
