% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{cols_rename}
\alias{cols_rename}
\title{Rename specific columns of a data.frame}
\usage{
cols_rename(df, old, new)
}
\arguments{
\item{df}{data.frame.}

\item{old}{Character vector.}

\item{new}{Character vector.}
}
\value{
df data.frame
}
\description{
Replace the column names of a data.frame entrywise for each element
a given vector (old) with the element of another vector (new)
}
\examples{
iris.renamed <- cols_rename(iris, "Sepal.Length", "Sepal Length")
iris.renamed2 <- cols_rename(
  iris,
  c("Sepal.Length", "Sepal.Width"),
  c("SLength", "SWidth")
)
}
