% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{colstartsw}
\alias{colstartsw}
\title{Searches Column Names Starting with an [Reg]Expression}
\usage{
colstartsw(regex = "", df)
}
\arguments{
\item{regex}{Character vector.}

\item{df}{data.frame.}
}
\description{
Prints out all Cols that start with a given string
no need to at .. to the regex. Helpful for very long data.frames
}
\examples{
colstartsw(iris, "Sepal")
}
