% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{corstars}
\alias{corstars}
\title{Corstars - Correaltions in Console}
\usage{
corstars(
  x,
  method = c("pearson", "spearman", "polychoric"),
  removeTriangle = c("upper", "lower", FALSE),
  rmDiag = c(TRUE, FALSE),
  rmLastCol = c(TRUE, FALSE),
  result = c("none", "html", "latex"),
  labels_rows = colnames(x),
  labels_cols = labels_rows[1:length(labels_rows)],
  sig.level = 0.05,
  nod = 2,
  caption = c("Correlation"),
  filename = ""
)
}
\arguments{
\item{x}{a matrix containing the data}

\item{method}{correlation method. "pearson", "spearman" or "polychoric" are supported}

\item{removeTriangle}{remove upper or lower triangle, or FALSE for not removing any triangle}

\item{rmDiag}{if one triangle of the matrix is removed, should the diagonal be kept = FALSE; or removed = TRUE}

\item{rmLastCol}{chose if the last column can be removed, to shorten the table if necessary, default = TRUE}

\item{result}{Print result in Console ("none"), generate HTML file ("html"), generate latex file ("latex")}

\item{labels_rows}{Labels for the rows (i.e., variable names). Length musst be same as number of variables}

\item{labels_cols}{Labels for columns. Length musst be same as number of variables - 1}

\item{sig.level}{Significance level (.1 or .05)}

\item{nod}{Integer. Number of Decimals. Default is nod = 2. In case of -1 a simple convention based
on sample size is applied for determination of number of decimal points.
See \code{\link{get_number_of_decimals}} or \code{?datscience::get_number_of_decimals}}

\item{caption}{Caption for the table}

\item{filename}{File name to save output to}
}
\value{
Correlation table in console or file
}
\description{
Creates a pretty console correlation table (by Dominik Vogel)
method : correlation method. "pearson"" or "spearman"" is supported
the results will be displayed in html or latex format.
labels_rows and labels_cols are character vectors for labeling rows and columns.
\href{https://rdrr.io/github/DominikVogel/vogelR/src/R/output.R}{Reference for the original code}.
Additionally added the option to investigate polychoric correlation
}
\examples{
\dontrun{
# Console output
corstars(mtcars,
  method = "pearson", removeTriangle = "upper", result = "none",
  caption = "Correlations",
  sig.level = 0.1,
  labels_rows = c(
    "(1) mpg", "(2) cyl", "(3) disp", "(4) hp",
    "(5) drat", "(6) wt", "(7) qsec", "(8) vs",
    "(9) am", "(10) gear",
    "(11) carb"
  ),
  labels_cols = 1:10
)

# HTML output
corstars(mtcars,
  method = "pearson", removeTriangle = "upper", result = "html",
  caption = "Correlations", filename = "corr.html",
  sig.level = 0.1,
  labels_rows = c(
    "(1) mpg", "(2) cyl", "(3) disp", "(4) hp",
    "(5) drat", "(6) wt", "(7) qsec", "(8) vs",
    "(9) am", "(10) gear",
    "(11) carb"
  ),
  labels_cols = 1:10
)
}
}
\seealso{
\code{\link{get_number_of_decimals}}
}
\author{
Dominik Vogel (Adapted by Bjoern Buedenbender)
}
