% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_table1.R
\name{flex_table1}
\alias{flex_table1}
\title{Creates a Descriptive Bivariate Table1 Ready for Publication}
\usage{
flex_table1(
  str_formula,
  data,
  correct = "bonf",
  num = NA,
  table_caption = NA,
  ref_correction = TRUE,
  include_teststat = TRUE,
  drop_unused_cats = TRUE,
  ...
)
}
\arguments{
\item{str_formula}{A string representing a formula, e.g., \code{"~ Sepal.Length + Sepal.Width | Species"}
used to construct the \code{\link[table1]{table1}}.}

\item{data}{The dataset containing the variables for the table1 call (all terms from the str_formula must be present)}

\item{correct}{Character, currently available are "bonf" for Bonferroni correction or "sidark" for Sidark correction.
Provide NA if you dont want any correction. Defaults to "bonf"}

\item{num}{Integer number of comparisons. If NA will be determined automatically, by the number of terms in the formula}

\item{table_caption}{Caption for the table, each element of the vector represents
a new line. The first line will be bold face. All additional lines are in italic.}

\item{ref_correction}{Boolean, default = TRUE, if TRUE corrected p-Values will be referenced in the foot note.}

\item{include_teststat}{Boolean, default = TRUE, if TRUE includes two additional columns in the table.
1) Test statistic (either t, f or X²) and 2) degrees of Freedom}

\item{drop_unused_cats}{Boolean, default = TRUE, if TRUE categories (i.e., factor levels) with 0 observations will be dropped.}

\item{...}{(Optional), Additional arguments that can be passed to \code{\link{format_flextable}}
(e.g., fontsize, font ...) or to \code{\link{serialNext}}}
}
\value{
A \code{\link[flextable]{flextable}} object with APA ready correlation table. If a filepath is provided
it also creates the respective file (e.g., a word .docx file)
}
\description{
A convenience function, that provides and easy wrapper for the two
   main enginges of the function \itemize{
        \item \code{\link[table1]{table1}} provides a nice API given a formula to create demographics tables.
        I basically just advanced the functionality of the p-value function, added the possibility
        to correct p-values with either Bonferroni or Sidark, and set some sensible defaults to achieve a nice look
        \item \code{\link[flextable]{flextable}} which gives all the power to format the table as you please
        (e.g., conditional formatting ->adding bold for p values below .05), adding italic headers or notes
        explaining what was done.
    }
    Really all credit should go to these two packages their developers, on full disclosure
    my function just provides an easy to use API or wrapper around their packages to get
    a beautiful publication ready bivariate comparison Table 1.
}
\details{
New feature since version (0.2.3) comparisons with (Welch) ANOVA for more than 2 Groups \cr
    \strong{On Fisher's Exact Test (FET) vs Pearson's χ²-test} \cr
    Newest feature (as of 07/22), according to an excellent post on cross-validated \insertCite{Harrell_cross_11}{datscience}
    the function refrains from using Fisher's exact test (FET) for categorical
    variables and only applies FET in the the rare case of cells with an expected
    cell frequencies do not exceed 1. This is due to the fact, that the FET can be
    extreme ressource intensive (and slow), and can have type I error rates less
    than the nominal level \insertCite{Crans2008}{datscience}
    Contemporary evidence suggests, that Pearson s χ²-test with the
    modification of \eqn{\frac{N-1}{N}}, nearly allways is more accurate than FET
    and generally recommended \insertCite{Lydersen2009}{datscience}. Thus
    in accordance we use the N-1 Pearson χ²-test proposed by (E.) Pearson and recommended
    as optimum test policy by \insertCite{Campbell2007}{datscience}.
}
\examples{

\dontrun{
  # Comparison of just two Groups
   str_formula <- "~ Sepal.Length + Sepal.Width +test | Species"
   data <- dplyr::filter(iris, Species \%in\% c("setosa", "versicolor"))
   data$test <- factor(rep(c("Female", "Male"), 50))
   table_caption <- c("Table 1", "A test on the Iris Data")
   flex_table1(str_formula, data = data, table_caption = table_caption)

  # Comparison of Multiple Groups (ANOVA)
   str_formula <- "~ Sepal.Length + Sepal.Width + Gender_example | Species"
   data <- dplyr::filter(iris, Species \%in\% c("setosa", "versicolor"))
   data <- iris
   data$Gender_example <- factor(rep(c("Female", "Male"), nrow(data)/2))
   table_caption <- c("Table 1", "A test on the Iris Data")
   flex_table1(str_formula, data = data, table_caption = table_caption)
   }

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{format_flextable}},
\code{\link[flextable]{flextable}}
\code{\link[table1]{table1}}
}
\author{
Bjoern Buedenbender,

Imports & Name Space
}
