% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{format_flextable}
\alias{format_flextable}
\title{Format a Flextable in Accordance with APA 7th Standards}
\usage{
format_flextable(
  ft,
  font = "Times New Roman",
  fontsize = 12,
  table_caption = c("Table x", "Some Description of the Table")
)
}
\arguments{
\item{ft}{A flextable object, to be formatted in accordance with APA. Required!}

\item{font}{Default is "Times New Roman", can be changed to your needs (e.g., "Arial")}

\item{fontsize}{Default is 12, bigger font size is not recommended.}

\item{table_caption}{Takes a character vector, with - the Title and # (e.g., "Table 1"); - and the descriptions; use capital case APA (e.g., Sociodemographic Characteristics of the Total Sample )}
}
\value{
An APA style formatted flextable object.
}
\description{
Formatting initially written by Remi Theriault, and advanced / wrapped in
a function by me
https://remi-theriault.com/blog_table.html
Reference on what APA style for tables constitutes can be obtained here
https://apastyle.apa.org/style-grammar-guidelines/tables-figures/tables
}
\author{
Björn Büdenbender / Remi Theriault
}
