% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICD_10.R
\name{get_ICD_10_cats}
\alias{get_ICD_10_cats}
\title{Get ICD-10 F-Codes as Character Vector}
\usage{
get_ICD_10_cats(lang = "eng")
}
\arguments{
\item{lang}{Language for the character vector, curr available in english and german (lang = "ger"), Default: "eng"}
}
\value{
Character Vector containing the 11 categroies for mental disorders (F Codes F01-F99)
}
\description{
Returns a character vector of length 11 for all ICD-10 F Categories (mental disorders)
}
\examples{
get_ICD_10_cats() # returns english ICD-10 Cats
get_ICD_10_cats("ger") # returns the german ones
}
\author{
Bjoern Buedenbender
}
