% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{getCIs}
\alias{getCIs}
\title{Get Boot Strapped CIs}
\usage{
getCIs(boot_obj)
}
\arguments{
\item{boot::boot}{object}
}
\value{
Confidence intervalls for all statistics
}
\description{
Determines the CI for a vector of statistics / mutliple stats.
As boot::boot.ci only returns CI for the first statistic in a boot_object
getCIs (given a boot::boot object), creates CI for all statistics.
Code by Ben Bolker, Check the original Source: https://stackoverflow.com/a/31818160/7318488
}
\author{
Ben Bolker
}
