% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_sample_means.R
\name{independent_sample_means}
\alias{independent_sample_means}
\title{Conduct a comparison for independent sample means - either t-test or one-way ANOVA}
\usage{
independent_sample_means(
  data,
  dv,
  iv,
  alternative = c("two.sided", "less", "greater"),
  stepwise = TRUE,
  verbose = TRUE,
  add = "jitter",
  fill = iv,
  palette = "jco",
  ...
)
}
\arguments{
\item{data}{The dataset containing the variables for the table1 call (all terms from the str_formula must be present)}

\item{dv}{The name of the dependend variable as character.}

\item{iv}{The name of the independend variable as character.}

\item{alternative}{In case of only two groups one can specify if a directed hypothesis should be tested. Default is "two.sided"}

\item{stepwise}{Boolean, default = TRUE, if TRUE the analysis is carried out in small steps, after each step (e.g. test for normality),
the output is printed is to the console and a user input is required. For a list of the steps see the description}

\item{verbose}{Boolean, default = TRUE, if TRUE the output of each step is printed to the console}

\item{add}{Additions to the boxplot, see also \link[ggpubr]{ggboxplot}. I primarily recommend to use either "none" or "jitter"}

\item{fill}{False for no filled colors}

\item{palette}{Color palette for the boxplot, see also \link[ggpubr]{ggboxplot}.}

\item{...}{(Optional), Additional arguments that can be passed to \code{\link[rstatix]{t_test}}
(e.g., fontsize, font ...) or to \code{\link{serialNext}}}
}
\value{
A list with all results of the check for the assumptions as well as the hypothesis test itself.
}
\description{
A convenience function, that provides and easy to use wrapper for a step-by-step comparison of means
from independent groups. The steps that are run for each analysis are:
\itemize{
\item \code{0. Determination of groups} For 2 groups conduct independent sample t-test. For more than 2 groups an
ANOVA is conducted instead
\item \code{1. Detect outliers and extreme values} including boxplots to identify those values. Utilizes \link[rstatix]{identify_outliers}
\item \code{2. Calculate descriptive statistics} for each groups, self-explanatory.
\item \code{3. Check noraml distribution (assumptions)} uses a formal test (Shapiro-Wilk) as well as a visual inspection method (qqplots).
Carefully judge this step, as the function has no default handling for non-normality. Options include check the literature if the test in your scenario is
robust for violations (see also central limit theorem and Glass, 1972), or switch to a non-parametric alternative.
\item \code{4. Check homogeneity of variances (assumption)} for this purpose a levene test will be conducted. A significant (by convention p < .05) result indicates
the H0 (of equal variances) must be disregarded. Depending on the result the student t-test with pooled variance / "normal" ANOVA
will be used. If there is heterogeneity of variances (i.e., unequal variances) the Welch correction of the degrees of freedom will be used instead.
\item \code{5. The test (t-Test or ANOVA)} will be conducted, including Post-Hoc Comparisons and a calculation of an effect size.
Additionally the ggpubr package is used together with rstatix to display a Boxplot including the result of the hypothesis test.
}
This function is a mere wrapper for the excellent tutorials provided by Alboukadel Kassambara over at Datanovia. See the original
\href{https://www.datanovia.com/en/lessons/t-test-assumptions/independent-t-test-assumptions/}{Guide for the t-test}
\href{https://www.datanovia.com/en/lessons/anova-in-r/}{and for the ANOVA}. It just packages all the steps in one function for
convenience
}
\examples{
t_test <- independent_sample_means(mtcars, dv = "disp", iv = "vs", add = "none")
ANOVA <- independent_sample_means(mtcars, dv = "disp", iv = "gear", add = "none")
}
\seealso{
\link[ggpubr]{ggboxplot}
\href{https://www.datanovia.com/en/lessons/t-test-assumptions/independent-t-test-assumptions/}{Guide for the t-test}
\href{https://www.datanovia.com/en/lessons/anova-in-r/}{and for the ANOVA}
}
\author{
Bjoern Buedenbender (adapted from Alboukadel Kassambara)
}
