% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{max.ina}
\alias{max.ina}
\title{Gives the Maximum Values Ignoring NAs}
\usage{
\method{max}{ina}(x)
}
\arguments{
\item{x}{Numeric Vector.}
}
\description{
A version of max that just ignores NA Values
}
