% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{my_apa}
\alias{my_apa}
\title{Creates APA Tables in Rmarkdown Templates}
\usage{
my_apa(df)
}
\arguments{
\item{df}{data.frame.}
}
\value{
Table
}
\description{
Given a data.frame creates an APA Table utilizing knitr::kable
This is also possible in Templates for RMarkdown
}
\examples{
my_apa(iris)
}
