% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{n1chisq.test}
\alias{n1chisq.test}
\title{Calculate a N-1 (E. Pearson) X²-Test}
\usage{
n1chisq.test(t)
}
\arguments{
\item{t}{A table / matrix}
}
\value{
A list with class "htest"
}
\description{
E. Pearson N-1 Correction to K. Pearson χ²-test Test, implementation
provided on stackoverflow and copied pasted \insertCite{wasserman}{datscience}
}
\examples{
\dontrun{
t <- matrix(c(1, 5, 3, 2), nrow = 2)
 n1chisq.test(t)
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[stats]{chisq.test}}
}
\author{
Wassermann
}
