% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_numberparse.R
\name{number_parse}
\alias{number_parse}
\title{number_parse}
\usage{
number_parse(s)
}
\arguments{
\item{s}{character or vector of characters, to be parsed to a number}
}
\description{
Takes one string or a vector of strings and converts them to numbers
Does retain the original dimensionality by rendering non convertable strings
to NA
}
