% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_cm2.R
\name{pretty_cm}
\alias{pretty_cm}
\title{pretty_cm - Pretty Caret Confusion Matrices}
\usage{
pretty_cm(
  cm,
  color_grad = c(alpha("yellowgreen", 0.4), alpha("springgreen3", 0.85)),
  midpoint = 50,
  hideZero = FALSE,
  ord = NA,
  plot = TRUE
)
}
\arguments{
\item{cm}{A \code{\link[caret]{confusionMatrix}} (mandatory parameter)}

\item{color_grad}{Pole of color gradient to use for the tiles, Default: c(alpha("yellowgreen", 0.4), alpha("springgreen3", 0.85))}

\item{midpoint}{Manually setting a middle point in percentage for the color scale, Default: 50}

\item{hideZero}{Hide tiles with 0 percentage, Default: FALSE}

\item{ord}{Order of the factor levels to display (if you want to change it manually for the plot), Default: NA}

\item{plot}{Shall the output also be plotted? Default: TRUE}
}
\value{
ggplot2 object - visualization of the confusion matrix.
}
\description{
Takes a caret confusion matrix and plots a nice visualization
}
\examples{
\dontrun{
if (interactive()) {
  # Creating random example data: prediction of neural network on content of animal pictures
  set.seed(23)
  pred <- factor(sample(c("dog", "cat"), 100, replace = TRUE))
  ref <- factor(sample(c("dog", "cat"), 100, replace = TRUE))
  cm <- caret::confusionMatrix(pred, ref)
  # Plotting of the caret confusion matrix
  pretty_cm(cm)
}
}
}
\seealso{
\code{\link[caret]{confusionMatrix}}
}
\author{
Björn Büdenbender
}
