% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_scree.R
\name{pretty_scree}
\alias{pretty_scree}
\title{Pretty Scree}
\usage{
pretty_scree(parallel, fa, quant = 0.95, custom_optimal = NA)
}
\arguments{
\item{parallel}{an parallel object returned by \code{psych::fa.parallel}}

\item{fa}{either "pc" or "fa" factor methods are allowed for the parallel analysis}

\item{quant}{default = .95 the quantile of the simulated values used to plot}

\item{custom_optimal}{default = NA, e.g., setting a marked dashed line to the location
determined by fa.parallel. Given 0 hides the line, all other value > 0 set the dashed
line to a custom location (e.g., to indicate a prefered number of factors determined by VSS or MAP)}
}
\value{
APA Ready Plot of Parallel Analyssis
}
\description{
Creates an Screeplot Including a Parallel Analysis (Horn), formatted according to APA
7th style. The original code was developed by John Sakaluk
Check the original source: \href{https://sakaluk.wordpress.com/2016/05/26/11-make-it-pretty-scree-plots-and-parallel-analysis-using-psych-and-ggplot2/#psych}{at his wordpress blog}.
}
\seealso{
\code{\link[psych]{fa.parallel}}
}
\author{
John Sakaluk (Wrapped in a Function and Advanced by Bjoern Buedenbender)
}
