% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{save_flextable}
\alias{save_flextable}
\title{Saves a Given Flextable Object}
\usage{
save_flextable(ft, filepath, overwrite = FALSE)
}
\arguments{
\item{ft}{A flextable object, to be formatted in accordance with APA. Required!}

\item{filepath}{Path and filename were the flextable object should be saved, options include the common filetypes
.docx (Word), .pptx (Powerpoint), .html (Webpage)}

\item{overwrite}{(Optional) Should the file, if already existing be overwritten}
}
\description{
Takes a flextable object (ft) and a filepath (path + filename + extension, e.g., "results/table1.docx")
And saves it. Primarily a utility function used by others in the package. Gives the option not to overwrite a
file by utilizing datscience::serialNext (from this package).
}
\author{
Bjoern Buedenbender
}
