% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{serialNext}
\alias{serialNext}
\title{Check if file exists, else append a integer number}
\usage{
serialNext(path, ignore_file_extension = T, n_digits = 3)
}
\arguments{
\item{path}{Character containing the path with fileextensions}

\item{ignore_file_extension}{Boolean currently only supports True, thus file extension will be
ignore in giving new filenames}

\item{n_digits}{Integer number of leadings 0 before the filename}
}
\value{
A new path and filename that currently does not exists
}
\description{
In cases you want to write to a file, however you do not want to override existing
files this helper function checks if a file exists, if it does it just appends an integer to the filename
Source originally by spacedman https://stackoverflow.com/a/25429755/7318488
}
\author{
spacedman stackoverflow (advanced by Björn Büdenbender)
}
