% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_swap.R
\name{spss_swap}
\alias{spss_swap}
\title{Swaps labels and col names from a .sav (SPSS) matrix}
\usage{
spss_swap(df)
}
\arguments{
\item{df}{tibble read in the \code{\link[haven]{read_sav}} from haven}
}
\value{
A tibble with more human readable names (old labels as names)
}
\description{
In many situations it can be useful to have the ability to replace the
uninformative variable names (colnames) from a SPSS matrix (e.g. SD02, SD08_02)
with their respective label attribute (e.g., "Gender/Sex" or "Age 15-25"). This
function utilizes the function \code{\link[janitor]{make_clean_names}} to convert the
labels from SPSS to proper variable names (e.g.,"gender_sex" or "age_15_25") and sets
them as new column names
}
\examples{
\dontrun{
datscience::spss_swap(df_sav)
}
}
\author{
Friedrich-Samuel Taubitz & Bjoern Buedenbender
}
