% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{util_rev_fac}
\alias{util_rev_fac}
\title{Reverses the Order of a Factor - Utility Function}
\usage{
util_rev_fac(f)
}
\arguments{
\item{f}{A factor}
}
\value{
Factor f with reversed order of the levels
}
\description{
Reverses the order of the levels of a factor.
}
\details{
Replacement for forcats::fct_rev(), in order to reduce the number
of dependencies to datscience package
}
\examples{
\dontrun{
if (interactive()) {
  f <- factor(letters[c(1:5)])
  (reversed_factor <- util_rev_fac(f))
}
}
}
\author{
Björn Büdenbender
}
