% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{apa_corrTable}
\alias{apa_corrTable}
\title{Create APA Publication Ready Correlation Table and Export to Word or Powerpoint}
\usage{
apa_corrTable(
  df,
  summarystats = c("mean", "sd"),
  method = c("pearson", "spearman", "polychoric"),
  rmDiag = FALSE,
  sig.level = 0.05,
  nod = c(2, -1),
  filepath = NA,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{df}{Data.frame, mandatory argument. Consider filtering before passing it
e.g. with dplyr::select() and dplyr::filter()}

\item{summarystats}{A vector with the summary stats to be included at the bottom
below the correlation. Default is c("mean","sd")
Options are one or all of c("mean","sd","median","range","min","max","skew",
"kurtosis","se","missing"). The option missing, adds missings per item as additional row
(accecpts both "missing" and "missings", spelling in table accordingly).
If NA is given, no summarystats will be added.}

\item{method}{Type of correlation. Options are currently: "pearson", "spearman" and "polychoric"}

\item{rmDiag}{Should the diagonal in the corr matrix kept (FALSE) or removed (TRUE)}

\item{sig.level}{How many stars per level of significance, options include
.05 .01 or .001}

\item{nod}{(Optional) Integer or Integer Vector. Number of Decimals.
In case of -1 a simple convention based on sample size is applied for determination
of number of decimal points. See ?datscience::get_number_of_decimals.
You can also provide an Integer vector, if you want different number of decimals
for the correlations and the summary stats. The first integer determines nod for
correlations, the second for summary stats. E.g., c(2,-1) would give 2 decimals
for correlations and apply the convention for summary stats. Default is nod = c(2,-1). \cr
See \code{\link{get_number_of_decimals}} or\code{?datscience::get_number_of_decimals}}

\item{filepath}{(Optional) Path and filename were the APA ready table should
be saved, options include the common filetypes .docx (Word), .pptx (Powerpoint),
.html (Webpage). Default is filepath = NA. If NA is given, no file will be saved.}

\item{overwrite}{(Optional) Boolean, default is FALSE. When overwrite is
FALSE and the files already exists, a serialized version of the filename
will be created (i.e., appending _001, or _002). Utilizes \code{\link{serialNext}}}

\item{...}{(Optional), Additional arguments that can be passed to \code{\link{format_flextable}}
(e.g., fontsize, font ...) or to \code{\link{serialNext}}}
}
\value{
A \code{\link[flextable]{flextable}} object with APA ready correlation table.
}
\description{
The Idea was inspired by a blog post of my colleague Remi Theriault
(see \href{https://remi-theriault.com/blog_table.html}{remi-theriault.com})
Which utilized the ability of \code{\link[flextable]{flextable}} to be able to get an
APA-style formatted table directly from R into .docx (word). One frequent use case
is to get a correlation table into word. For the creation of the correlation table
this function uses the \code{\link{corstars}} function from this package, which
resolves around the code by
\href{https://rdrr.io/github/DominikVogel/vogelR/src/R/output.R}{Dominik Vogel}.\cr
I built on those idea and created a function that creates a pretty correlation table
together with the summary stats of your choice, and returns a flextable, which
can be easily exported to word (see also  \code{\link{save_flextable}})\cr
\strong{Please note:} This function only considers numeric variables (cols). Other
datatypes will be dropped from the data.frame df.
}
\examples{
apa_corrTable(mtcars, table_caption=c("Table 2","Correlations in the mtcars Data Set"))

}
\seealso{
\code{\link{get_number_of_decimals}}, \code{\link{format_flextable}},
\code{\link{serialNext}}
}
\author{
Bjoern Buedenbender (Inspired by Remi Theriault)
}
