% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{booted_eigenvalues}
\alias{booted_eigenvalues}
\title{Booted Eigenvalues}
\usage{
booted_eigenvalues(
  df,
  iterations = 1000,
  cor = "pearson",
  fa = "pc",
  fm = "minres"
)
}
\arguments{
\item{df}{A data.frame}

\item{iterations}{number of resamples for the bootstrap}

\item{cor}{either "pearson" or "poly" for polychoric correlations,
defaults to "pearson"}

\item{fa}{either "pc" for prinicipal component or
"fa" for [common] factor analysis, defaults to "pc"}

\item{fm}{factor method to use, irrelevant for pca.
For available factor methods check psych::fa for more details, defaults to minres}
}
\value{
A boot object (\code{boot::boot()}),
that contains SE for all Eigenvalues in DF, can be passed to \code{\link{getCIs}}  to create Confidence Intervals
}
\description{
Uses \code{\link[boot]{boot}}  to bootstrap Eigenvalues in factor analyses. \cr
Eigenvalues can be extracted for PCA as well as EFA \cr
Different Factor Methods are available for more details, see \code{\link[psych]{fa}}
Allows for extraction of Eigenvalues based on Pearson
as well as polychoric correlation matrices. These bootstrapped Eigenvalues can
serve as a foundation to construct a scree plot with confidence intervals around
the observed eigenvalues.
}
\seealso{
\code{\link[boot]{boot}}, \code{\link[psych]{fa}}, \cr
\code{\link{getCIs}} \code{\link{add_ci_2plot}}
}
\author{
Bjoern Buedenbender
}
