% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_flextable.R
\name{format_flextable}
\alias{format_flextable}
\title{Format a Flextable in Accordance with APA 7th Standards}
\usage{
format_flextable(
  ft,
  font = "Times New Roman",
  fontsize = 12,
  table_caption = c("Table x", "Some Description of the Table"),
  table_note = NA,
  italize_stats = TRUE,
  linespacing = 1.25,
  ...
)
}
\arguments{
\item{ft}{A flextable object (prefered) or data.frame, to be formatted in accordance with APA. Required!}

\item{font}{Default is "Times New Roman", can be changed to your needs (e.g., "Arial")}

\item{fontsize}{Default is 12, bigger font size is not recommended.}

\item{table_caption}{Takes a character vector. Note: when provided NA no caption will
be placed! Recommend are the following elements
\itemize{
\item Table + number (e.g., "\strong{Table 1}")
\item The description, use APA capital case
(e.g., "\emph{Sociodemographic Characteristics of the Total Sample}")
}}

\item{table_note}{Takes a character vector. Will automatically prepend Note if not present.
Default is NA (no table note). e.g., is c("Explanation of Abbreviations",
"* p < .05. ** p < .01. *** p < .001").
Please note that you will need to manually set the formatting (i.e., setting italic for p values).}

\item{italize_stats}{Boolean, default is TRUE. Searches the header for common statistic symbols (e.g. p N SD M) and
sets font to italic}

\item{linespacing}{Points of Linespacing defaults to 1.25, APA would be 1.5 or 2}

\item{...}{(Optional) Additional Parameters, not utiliezd in this function, enables passing from passing of previous functions}
}
\value{
An APA style formatted flextable object.
}
\description{
This function takes a flextable object, and applies APA 7th
standard style formatting to it. \cr
The initial idea for this function stems from
\href{https://remi-theriault.com/blog_table.html}{Remi Theriault's Blog}. \cr
Reference on what APA style for tables constitutes can be obtained on the
\href{https://apastyle.apa.org/style-grammar-guidelines/tables-figures/tables}{apastyle.apa.org website}.
See also the very neat packages \code{\link[flextable]{flextable}} and \code{\link[officer]{read_docx}},
by David Gohel and Colleagues, which enables us to get our publication ready tables directly
into MS Word.
}
\seealso{
\code{\link[flextable]{flextable}}
}
\author{
Bjoern Buedenbender / Remi Theriault
}
