% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{get_number_of_decimals}
\alias{get_number_of_decimals}
\title{Determine Number of Decimals by a Simple Convention (Based on N)}
\usage{
get_number_of_decimals(n)
}
\arguments{
\item{n}{Number of observations (e.g., participants in a study)}
}
\value{
Integer the appropriate number of decimals, based on sample size (n)
}
\description{
A simple \href{https://support.jmir.org/hc/en-us/articles/360019690851-Guidelines-for-Reporting-Statistics}{convention}
for the determination of the appropriate number of decimals
will be applied, e.g.,  \emph{n}
\itemize{
\item < 100: no decimal
\item < 1000: 1 decimals
\item > 1001: 2 decimals
}
}
\examples{
get_number_of_decimals(n = 153)

}
\author{
Bjoern Buedenbender
}
