% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{getCIs}
\alias{getCIs}
\title{Get Boot Strapped CIs}
\usage{
getCIs(boot_obj)
}
\arguments{
\item{boot_obj}{boot::boot object}
}
\value{
A data.frame with Confidence intervalls for
all statistics, as well as the observed value
}
\description{
Determines the CI for a vector of statistics / multiple stats.
As \code{\link[boot]{boot.ci}} only returns CI for the first statistic in a boot_object
getCIs (given a \code{\link[boot]{boot}} object), creates CI for all statistics.
Code by Ben Bolker, Check the
original Source: \href{https://stackoverflow.com/a/31818160/7318488}{@stackoverflow}
}
\seealso{
\code{\link[boot]{boot.ci}}, \code{\link[boot]{boot}}
}
\author{
Ben Bolker
}
