% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{pretty_scree}
\alias{pretty_scree}
\title{Pretty Scree}
\usage{
pretty_scree(parallel, fa, quant = 0.95)
}
\arguments{
\item{parallel}{an parallel object returned by psych::fa.parallel}

\item{fa}{either "pc" or "fa" the factor methods used for the parallel analysis}

\item{quant}{default = .95 the quantile of the simulated values used to plot}
}
\value{
APA Ready Plot of Parallel Analyssis
}
\description{
Creates an APA Plot of Parallel Analysis (Horn)
Check the original Source: https://sakaluk.wordpress.com/2016/05/26/11-make-it-pretty-scree-plots-and-parallel-analysis-using-psych-and-ggplot2/#psych
}
\author{
John Sakaluk (Wrapped in a Function by Björn Büdenbender)
}
