% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datscience.R
\name{Rcitation_appendix}
\alias{Rcitation_appendix}
\title{R Package Citations}
\usage{
Rcitation_appendix(
  outdirectory = "Appendix",
  filename = "Appendix - R Packages.docx",
  overwrite = FALSE,
  table_caption = c("Table A1", "All R-Packages Utilized and Dependencies"),
  ...
)
}
\arguments{
\item{outdirectory}{(Optional) Character vector for the output directory (for the two files, .bib and .csv). Default is "Appendix/" in the current working directory (see with getwd())}

\item{filename}{(Optional) Character vector. Custom name for the formatted APA style table, that should end with .docx. A docx file will be provided. If NA, a csv file of the table will be saved.}

\item{overwrite}{(Optional) Boolean, default is FALSE. When overwrite is FALSE and the files already exists, a serialized version of the filename will be created (i.e., appending _001, or _002)}

\item{table_caption}{(Optional) Takes a character vector, with - the Title and # (e.g., "Table A1"); - and the descriptions; use capital case APA (e.g., "Complete List of All R-Packages Utilized and Dependencies" )}

\item{...}{(Optional), Additional arguments that can be passed to format_flextable (e.g., fontsize), please refer to ?datscience::format_flextable}
}
\value{
- Creates the directory Appendix/ (if not otherwise specified)
- In the "Appendix/" dir a .bib file is created for the citation
- Either creates an APA 7th style table in a .docx or a .csv table
- Additionally returns the flextable::flextable object with the APA 7th style table.
}
\description{
Function to create a package table with the following columns: Packagename | Version | Maintainer | Citation
}
\author{
Bjoern Buedenbender
}
