% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialNext.R
\name{serialNext}
\alias{serialNext}
\title{Check if file exists, else append a integer number}
\usage{
serialNext(path, n_digits = 3, maxruns = 500, ...)
}
\arguments{
\item{path}{Character containing the path with fileextensions}

\item{n_digits}{(Optional) Integer, number of leadings 0 before the filename. Default is n_digits = 3}

\item{maxruns}{(Optional) Integer, Default is maxruns = 500.}

\item{...}{(Optional) Additional Parameters, not utiliezd in this function, enables passing from passing of previous functions}
}
\value{
Character, a filename that currently does not exists, with indices.
}
\description{
In cases you want to write to a file which you do not want to override if existing.
This helper function checks if a file exists, if it does it just appends an integer to the filename
Source originally by user spacedman \href{https://stackoverflow.com/a/25429755/7318488}{on stackoverflow}.
}
\author{
spacedman stackoverflow (advanced by Bjoern Buedenbender)
}
